apps/api/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Database
asyncpg==0.29.0
aiomysql==0.2.0
sqlalchemy[asyncio]==2.0.30
psycopg2-binary==2.9.9

# LLM providers
openai==1.35.0
anthropic==0.30.0

# Vector DB
qdrant-client==1.9.2

# Background jobs
celery[redis]==5.4.0
redis==5.0.7

# SQL validation
sqlglot==25.5.1

# Utilities
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4
httpx==0.27.0
cryptography==42.0.8